refactor: move Notion and Prefect clients to app/services/ and update imports
