+++ app/integrations/clients/prefect/__init__.py
@@ -1,9 +1,12 @@
-"""Prefect API client package."""
+"""Compatibility shim for moved Prefect client."""
+import warnings
 
-from app.integrations.clients.prefect.client import (
-    PrefectClient,
-    PrefectConfig,
-    make_prefect_client,
+from app.services.prefect import PrefectClient, PrefectConfig, make_prefect_client
+
+warnings.warn(
+    "app.integrations.clients.prefect is deprecated. Use app.services.prefect instead.",
+    DeprecationWarning,
+    stacklevel=2,
 )
 
 __all__ = ["PrefectClient", "PrefectConfig", "make_prefect_client"]
