+++ app/services/notion/__init__.py
@@ -1,0 +1,4 @@
+"""Notion service adapter."""
+from .client import NotionClient, NotionConfig
+
+__all__ = ["NotionClient", "NotionConfig"]
