+++ app/services/prefect/__init__.py
@@ -1,0 +1,9 @@
+"""Prefect API client package."""
+
+from .client import (
+    PrefectClient,
+    PrefectConfig,
+    make_prefect_client,
+)
+
+__all__ = ["PrefectClient", "PrefectConfig", "make_prefect_client"]
