+++ app/tools/PrefectWorkerHealthTool/__init__.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from app.integrations.clients.prefect import make_prefect_client
+from app.services.prefect import make_prefect_client
 from app.tools.base import BaseTool
 
 _UNHEALTHY_WORKER_STATUSES = {"OFFLINE", "UNHEALTHY"}
